Refactor: use asset cache
